requirements.txt
# Core runtime
PySimpleGUI
Pillow
imageio
numpy

# Test tooling
pytest
